+++ backend/src/v1/schema/common.rs
@@ -4,6 +4,6 @@ use ts_rs::TS;
 #[derive(TS, Debug, Serialize, Deserialize, Clone)]
 #[ts(export)]
 pub struct Pagination {
-    pub page: u64,
-    pub per_page: u64,
+    pub page: u32,
+    pub per_page: u32,
 }
